+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+# 1.5.2
+- Relaunch: the memory view of the first session is now correctly closed before re-launching.
+- Commands: "-view" is now working with all commands.
+- Commands:
+	- Renamed: "md" changed to "mv".
+	- "md" used to do a memory dump to console.
+	- "ms" used to save memory contents to a file.
+
 # 1.5.1
 - Packaged for beta testing.